requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.3
packaging==21.3
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.20.1
psutil==5.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scipy==1.9.2
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sudoku-solver==0.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
urllib3==1.26.12
wandb==0.13.3
